added gateways + refactor
